chore: modularizing
