+++ src/entry.html
@@ -4,8 +4,8 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Quick Notes</title>
-  <link rel="stylesheet" href="stylesheet.css">
-  <script type="module" src="src/js/entry.js"></script>
+  <link rel="stylesheet" src="styles/stylesheet.css">
+  <script type="module" src="js/entry.js"></script>
 </head>
 <body>
   <div class="container">
