+++ src/js/entry.js
@@ -3,7 +3,7 @@ import { saveNote } from './saveNotes.js';
 
 document.addEventListener("DOMContentLoaded", () => {
   const notesList = document.getElementById("notesList");
-  const saveButton = document.getElementById("saveButton");
+  const saveButton = document.getElementById("saveBtn");
   const noteInput = document.getElementById("noteInput");
 
   // Initial display of notes
@@ -12,6 +12,7 @@ document.addEventListener("DOMContentLoaded", () => {
   // Save new note when button is clicked
   saveButton.addEventListener("click", async () => {
     const newNote = noteInput.value;
+    console.log(newNote);
     if (newNote) {
       await saveNote(newNote, notesList);
       noteInput.value = ""; // Clear input field after saving
